Move terminal UI into dedicated module
