+++ app/modules/AgentHubCore/Sources/AgentHub/Services/DevServerManager.swift
@@ -7,6 +7,7 @@
 //
 
 import Foundation
+import AgentHubTerminalUI
 import Storybook
 
 // MARK: - DevServerManager
