+++ app/modules/AgentHubCore/Sources/AgentHub/UI/ResizableCardContainer.swift
@@ -1,4 +1,5 @@
 import AppKit
+import AgentHubTerminalUI
 import SwiftUI
 
 struct ResizableCardMetrics {
@@ -25,33 +26,6 @@ enum ResizableCardHandlePlacement {
 enum ResizableCardHandleStyle {
   case line
   case grip
-}
-
-final class ResizeInteractionSuppression {
-  static let shared = ResizeInteractionSuppression()
-
-  private let lock = NSLock()
-  private var activeResizeCount = 0
-  private var suppressSelectionUntil: TimeInterval = 0
-
-  func beginResize() {
-    lock.lock()
-    activeResizeCount += 1
-    lock.unlock()
-  }
-
-  func endResize() {
-    lock.lock()
-    activeResizeCount = max(0, activeResizeCount - 1)
-    suppressSelectionUntil = max(suppressSelectionUntil, ProcessInfo.processInfo.systemUptime + 0.2)
-    lock.unlock()
-  }
-
-  var shouldSuppressSelection: Bool {
-    lock.lock()
-    defer { lock.unlock() }
-    return activeResizeCount > 0 || ProcessInfo.processInfo.systemUptime < suppressSelectionUntil
-  }
 }
 
 struct ResizableCardContainer<Content: View>: View {
