+++ app/modules/AgentHubCore/Sources/AgentHubTerminalUI/EmbeddedTerminalBackend.swift
@@ -17,7 +17,7 @@ public enum EmbeddedTerminalBackend: Int, CaseIterable, Sendable {
   }
 
   public static var storedPreference: EmbeddedTerminalBackend {
-    let rawValue = UserDefaults.standard.object(forKey: AgentHubDefaults.terminalBackend) as? Int
+    let rawValue = UserDefaults.standard.object(forKey: TerminalUserDefaultsKeys.terminalBackend) as? Int
       ?? EmbeddedTerminalBackend.regular.rawValue
     return EmbeddedTerminalBackend(rawValue: rawValue) ?? .regular
   }
