+++ app/modules/AgentHubCore/Sources/AgentHubTerminalUI/EmbeddedTerminalHostView.swift
