+++ app/modules/AgentHubCore/Sources/AgentHubTerminalUI/EmbeddedTerminalSurface.swift
@@ -14,29 +14,33 @@ public protocol EmbeddedTerminalSurface: AnyObject {
   var onRequestShowEditor: (() -> Void)? { get set }
   var consumeQueuedWebPreviewContextOnSubmit: (() -> String?)? { get set }
   var onWorkspaceChanged: ((TerminalWorkspaceSnapshot) -> Void)? { get set }
+  var onOpenFile: ((String, Int?) -> Void)? { get set }
 
-  func updateContext(terminalSessionKey: String?, sessionViewModel: CLISessionsViewModel?)
+  func updateContext(terminalSessionKey: String?)
   func configure(
-    sessionId: String?,
+    launch: Result<EmbeddedTerminalLaunch, EmbeddedTerminalLaunchError>,
     projectPath: String,
-    cliConfiguration: CLICommandConfiguration,
-    initialPrompt: String?,
     initialInputText: String?,
-    isDark: Bool,
-    dangerouslySkipPermissions: Bool,
-    permissionModePlan: Bool,
-    worktreeName: String?,
-    metadataStore: SessionMetadataStore?
+    isDark: Bool
   )
-  func configureShell(projectPath: String, isDark: Bool, shellPath: String?)
-  func restart(sessionId: String?, projectPath: String, cliConfiguration: CLICommandConfiguration)
+  func configureShell(
+    launch: EmbeddedTerminalLaunch,
+    projectPath: String,
+    isDark: Bool
+  )
+  func restart(launch: Result<EmbeddedTerminalLaunch, EmbeddedTerminalLaunchError>, projectPath: String)
   func terminateProcess()
   func resetPromptDeliveryFlag()
   func sendPromptIfNeeded(_ prompt: String)
   func submitPromptImmediately(_ prompt: String) -> Bool
   func typeText(_ text: String)
   func typeInitialTextIfNeeded(_ text: String)
-  func syncAppearance(isDark: Bool, fontSize: CGFloat, fontFamily: String, theme: RuntimeTheme?)
+  func syncAppearance(
+    isDark: Bool,
+    fontSize: CGFloat,
+    fontFamily: String,
+    theme: TerminalAppearanceTheme?
+  )
   func focus()
   func captureWorkspaceSnapshot() -> TerminalWorkspaceSnapshot?
   func restoreWorkspaceSnapshot(_ snapshot: TerminalWorkspaceSnapshot)
@@ -44,6 +48,11 @@ public protocol EmbeddedTerminalSurface: AnyObject {
 
 public extension EmbeddedTerminalSurface {
   var onWorkspaceChanged: ((TerminalWorkspaceSnapshot) -> Void)? {
+    get { nil }
+    set {}
+  }
+
+  var onOpenFile: ((String, Int?) -> Void)? {
     get { nil }
     set {}
   }
@@ -88,9 +97,8 @@ public struct DefaultEmbeddedTerminalSurfaceFactory: EmbeddedTerminalSurfaceFact
 extension TerminalContainerView: EmbeddedTerminalSurface {
   public var view: NSView { self }
 
-  public func updateContext(terminalSessionKey: String?, sessionViewModel: CLISessionsViewModel?) {
+  public func updateContext(terminalSessionKey: String?) {
     self.terminalSessionKey = terminalSessionKey
-    self.sessionViewModel = sessionViewModel
   }
 
   public func focus() {
