+++ app/modules/AgentHubCore/Sources/AgentHubTerminalUI/RegularTerminalPaneHeader.swift
