+++ app/modules/AgentHubCore/Sources/AgentHubTerminalUI/TerminalPromptSubmissionPayload.swift
