+++ app/modules/AgentHubCore/Sources/AgentHubTerminalUI/TerminalSubmitInterception.swift
