+++ app/modules/AgentHubCore/Sources/AgentHubTerminalUI/TerminalUILogger.swift
@@ -1,0 +1,10 @@
+//
+//  TerminalUILogger.swift
+//  AgentHub
+//
+
+import OSLog
+
+public enum TerminalUILogger {
+  public static let terminal = Logger(subsystem: "com.agenthub", category: "Terminal")
+}
