+++ app/modules/AgentHubCore/Sources/AgentHubTerminalUI/TerminalWorkspaceSnapshot.swift
