+++ app/modules/AgentHubCore/Tests/AgentHubTests/NewlineShortcutTests.swift
@@ -1,5 +1,6 @@
 import Testing
 @testable import AgentHubCore
+@testable import AgentHubTerminalUI
 
 @Suite("NewlineShortcut")
 struct NewlineShortcutTests {
